M3-5603: Clean up GooglePay code (#8070)
* clean up gpay code

* delete old CreditCardInfo component

* rename file to GooglePayProvider

* address todo for displaying API error
